Creo que ya va auqnue me sigue fallando el spring

+++ src/miHexagonalui_driver_adapter/JpaBookRepository.java
@@ -4,7 +4,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import miHexagonal.core.domain.Book;
+import miHexagonal.core.domain.BookDoesNotExistException;
 import miHexagonal.core.driven_ports.BookRepository;
+import miHexagonaldb_driven_adapter.jparepository.HAAJpaRepository;
 
 @Component
 public class JpaBookRepository implements BookRepository {
@@ -12,7 +14,7 @@ public class JpaBookRepository implements BookRepository {
 @Autowired //enlaza repositorio y servicio
 private HAAJpaRepository haaJpaRepository;
     @Override
-public Book findById(Long id) {
+public Book findById(int id) {
 Optional<BookEntity> bookEntityOptional = haaJpaRepository.findById(id);
 BookEntity bookEntity = bookEntityOptional.orElseThrow(BookDoesNotExistException::new);
 return bookEntity.toBook();
